Change histogram bin count to long type (and introduce a brief deserialization test to make sure I got it right)

+++ src/test/java/io/inscopemetrics/kairosdb/integration/AggregationIT.java
@@ -225,7 +225,7 @@ public final class AggregationIT {
     public void testMergeAggregatorMulti() throws IOException, JSONException {
         final List<Double> numbers = Lists.newArrayList();
         for (Histogram histogram : MULTI_HIST_TEST_DATA) {
-            for (final Map.Entry<Double, Integer> entry : histogram.getBins().entrySet()) {
+            for (final Map.Entry<Double, Long> entry : histogram.getBins().entrySet()) {
                 for (int count = 0; count < entry.getValue(); count++) {
                     numbers.add(entry.getKey());
                 }
